#65 - Implementar a opção de desconto nos produtos #66 - Implementar a opção de acréscimo nos produtos #67 - Implementar a opção de alterar preço dos materiais

+++ branches/0.4.0/app/src/main/java/br/com/informsistemas/forcadevenda/controller/adapter/MovimentoAdapter.java
@@ -59,7 +59,7 @@ public class MovimentoAdapter extends RecyclerView.Adapter<MovimentoAdapter.MyVi
 
         myViewHolder.txtCodigoParceiro.setText(p.codigoparceiro);
         myViewHolder.txtDescricao.setText(p.descricao);
-        myViewHolder.txtTotalLiquido.setText(Misc.formatMoeda(fList.get(position).totalliquido));
+        myViewHolder.txtTotalLiquido.setText(Misc.formatMoeda(fList.get(position).totalliquido.floatValue()));
 
         for (int i : selectedIds){
             if (fList.get(position).id == i){
